Refactor: Convert JS to TS
